+++ applications/services/cli_f20/application.fam
@@ -1,7 +1,6 @@
 App(
-    appid="cli",
+    appid="cli_commands_target_specific",
     apptype=FlipperAppType.STARTUP,
     entry_point="cli_on_system_start",
-    cdefines=["SRV_CLI"],
     targets=["f20", "f21"],
 )
