+++ applications/services/cli_f20/cli_command_factory_reset.c
