+++ applications/services/cli_f20/cli_command_status_lights.c
