--- applications/services/cli_f64/cli_commands.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#define RECORD_CLI "cli"
