--- applications/services/cli_f64/cli_main_shell.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <cli/cli_command.h>
-
-void cli_main_motd(void* context);
