--- applications/services/cli_f64/cli_uart.h
@@ -1,18 +1,0 @@
-/**
- * @file cli_uart.h
- * VCP HAL API
- */
-
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-typedef struct CliSession CliSession;
-
-extern CliSession cli_uart;
-
-#ifdef __cplusplus
-}
-#endif
