+++ lib/cli/cli_commands.h
